network-microservice/requirements.txt
fastapi
uvicorn
nornir
nornir-napalm
nornir-netmiko
nornir-utils
pydantic
pyyaml